[level 2] Title: 시소 짝꿍, Time: 0.24 ms, Memory: 75.2 MB -BaekjoonHub

+++ 프로그래머스/2/152996. 시소 짝꿍/시소 짝꿍.java
@@ -1,14 +1,14 @@
 class Solution {
     public long solution(int[] weights) {
         long answer = 0;
-        int[] DAT = new int[1001];
+        long[] DAT = new long[1001];
         for (int weight: weights) {
             DAT[weight]++;
         }
         for (int i = 100; i <= 1000; i++) {
-            int cnt = DAT[i];
+            long cnt = DAT[i];
             if (cnt >= 2) {
-                answer += ((long)cnt * (cnt - 1)) / 2;
+                answer += (cnt * (cnt - 1)) / 2;
             }
         }
         
@@ -16,7 +16,7 @@ class Solution {
         for (int i = 100; i <= 1000; i++) {
             for (int[] a: arr) {
                 if (i % a[0] != 0 || i / a[0] * a[1] > 1000) continue;
-                answer += (long)DAT[i] * DAT[i / a[0] * a[1]];
+                answer += DAT[i] * DAT[i / a[0] * a[1]];
             }
         }
